fix: fix deleting a winner

+++ async-race/src/pages/garage/garage.ts
@@ -91,7 +91,12 @@ export class Garage extends BasePage {
   async removeCar(id: number): Promise<void> {
     this.toggleAllButtonsMode();
     await this.asyncRaceApi.deleteCar(id, this.pageName);
-    await this.asyncRaceApi.deleteCar(id, 'winners');
+    const currentWinner = (await this.asyncRaceApi.getWinners()).find(
+      (elem) => elem.id === id,
+    );
+    if (currentWinner) {
+      await this.asyncRaceApi.deleteCar(id, 'winners');
+    }
     this.refreshTotalCount();
     await this.refreshCurrentPage();
     this.toggleAllButtonsMode(false);
